+++ fingerprint/config.js
@@ -1,4 +1,6 @@
-module.exports = function fingerprintConfig ({static=[]}) {
+module.exports = function fingerprintConfig (params) {
+  let { static=[] } = params
+
   // Maybe enable fingerprint
   let fingerprint = false
   if (static.some(s => {
@@ -8,6 +10,7 @@ module.exports = function fingerprintConfig ({static=[]}) {
   })) {
     fingerprint = true
   }
+
   // Allow apps and frameworks to handle their own fingerprinting
   if (static.some(s => {
     if (!s[0]) return false
@@ -19,8 +22,12 @@ module.exports = function fingerprintConfig ({static=[]}) {
 
   // Collect any strings to match against for ignore
   let ignore = static.find(s => s['ignore'])
-  if (ignore) {ignore = Object.getOwnPropertyNames(ignore.ignore)}
+  if (ignore) {
+    ignore = process.env.DEPRECATED
+      ? Object.getOwnPropertyNames(ignore.ignore)
+      : ignore.ignore
+  }
   else ignore = []
 
-  return {fingerprint, ignore}
+  return { fingerprint, ignore }
 }